+++ Thunderstore/CHANGELOG.md
@@ -2,6 +2,7 @@
 - small code clean-up
 - fixed `give item` command `food` showing up when eating
 - added API for other plugin's to add commands
+- added death quips, configurable, name of file must not be changed or it won't be able to update
 
 # 1.0.0
 - Initial release